separation of static and dynamic data
